Added logging module, Express web framework, Bootstrap, jquery and lib folder for common files.
Wired up Express to support responsive layout design and mobile UI using
Bootstrap and jquery. Added Winston to handle logging.
The registry of things array is moved to the libs folder. The idea
behind this to decouple the resources so they are available for the
application files such as the web server and the route files without the
need of passing it as a reference. This implementation also support
persistent store of the things list, so multiple load balanced web
servers can connect to the same list in the future.